move validation schemas to separate file
